+++ k2.go
@@ -4,6 +4,7 @@ import (
 	"encoding/json"
 	"github.com/julienschmidt/httprouter"
 	"net/http"
+	"strings"
 )
 
 type K2ResponseData struct {
@@ -16,50 +17,60 @@ type SaveToK2Data struct {
 	Data string
 }
 
-func saveCode(writer http.ResponseWriter, request *http.Request, _ httprouter.Params) {
-	logInfo("Save Code", "Saving code called")
+func saveDataToK2(writer http.ResponseWriter, request *http.Request, _ httprouter.Params) {
+	ipAddress := strings.Split(request.Host, ":")
+	deviceName := devicesMap[ipAddress[0]]
+	if len(deviceName) == 0 {
+		deviceName = ipAddress[0]
+	}
+	logInfo(deviceName, "Saving data to K2 called")
 	var data SaveToK2Data
 	err := json.NewDecoder(request.Body).Decode(&data)
 	if err != nil {
-		logError("MAIN", err.Error())
+		logError(deviceName, "Error parsing data from page: "+err.Error())
 		return
 	}
-	logInfo("Save Code", "Saving code "+data.Data)
+	logInfo(deviceName, "Data: "+data.Data)
 	var responseData K2ResponseData
 	responseData.Data = "ok"
 	writer.Header().Set("Content-Type", "application/json")
 	_ = json.NewEncoder(writer).Encode(responseData)
 	//TODO: SaveCodeToK2
-	logInfo("Save Code", "Saving code finished")
+	logInfo(deviceName, "Saving data to K2 finished, everything ok")
 
 }
 
-func checkOrder(writer http.ResponseWriter, request *http.Request, _ httprouter.Params) {
-	logInfo("Check Order", "Checking order called")
+func checkOrderInK2(writer http.ResponseWriter, request *http.Request, _ httprouter.Params) {
+	ipAddress := strings.Split(request.Host, ":")
+	deviceName := devicesMap[ipAddress[0]]
+	if len(deviceName) == 0 {
+		deviceName = ipAddress[0]
+	}
+	logInfo(deviceName, "Check order in K2 called")
 	var data SaveToK2Data
 	err := json.NewDecoder(request.Body).Decode(&data)
 	if err != nil {
-		logError("MAIN", err.Error())
+		logError(deviceName, "Error parsing data from page: "+err.Error())
 		return
 	}
-	orderIsInSystem := checkOrderInSystem(data.Data)
+	logInfo(deviceName, "Data: "+data.Data)
+	orderIsInSystem := checkOrderInSystem(deviceName, data.Data)
 	var responseData K2ResponseData
 	if !orderIsInSystem {
 		responseData.Data = "nok"
 		writer.Header().Set("Content-Type", "application/json")
 		_ = json.NewEncoder(writer).Encode(responseData)
-		logInfo("Check Order", "Checking order finished")
+		logInfo(deviceName, "Checking order in K2 finished, order not in system ")
 		return
 	}
 	responseData.Data = "ok"
 	writer.Header().Set("Content-Type", "application/json")
 	_ = json.NewEncoder(writer).Encode(responseData)
-	logInfo("Check Order", "Checking order finished")
+	logInfo(deviceName, "Check order in K2 finished, everything ok")
 }
 
-func checkOrderInSystem(order string) bool {
-	logInfo("Check Order In System", "Checking order "+order)
-	logInfo("Check Order In System", "Order checked ")
+func checkOrderInSystem(deviceName string, order string) bool {
+	logInfo(deviceName, "Checking order "+order)
 	//TODO: check order in K2
 	if order == "12345" {
 		return true
@@ -67,19 +78,24 @@ func checkOrderInSystem(order string) bool {
 	return false
 }
 
-func getK2Pcs(writer http.ResponseWriter, request *http.Request, _ httprouter.Params) {
-	logInfo("Get K2 Pcs", "Getting K2 pcs called")
+func getPcsFromK2(writer http.ResponseWriter, request *http.Request, _ httprouter.Params) {
+	ipAddress := strings.Split(request.Host, ":")
+	deviceName := devicesMap[ipAddress[0]]
+	if len(deviceName) == 0 {
+		deviceName = ipAddress[0]
+	}
+	logInfo(deviceName, "Get pcs from K2 called")
 	var data SaveToK2Data
 	err := json.NewDecoder(request.Body).Decode(&data)
 	if err != nil {
-		logError("Get K2 Pcs", err.Error())
+		logError(deviceName, "Error parsing data from page: "+err.Error())
 		return
 	}
-	logInfo("Get K2 Pcs", "Getting K2 pcs for order "+data.Data)
+	logInfo(deviceName, "Data: "+data.Data)
 	//TODO: Get Pcs from K2
 	var responseData K2ResponseData
 	responseData.Data = "23"
 	writer.Header().Set("Content-Type", "application/json")
 	_ = json.NewEncoder(writer).Encode(responseData)
-	logInfo("Get K2 Pcs", "Getting K2 pcs finished")
+	logInfo(deviceName, "Get pcs from K2 finished, everything ok")
 }